fixed lookup string "digitsStr" to match excel
Thanks to Peter Wilks for reporting a mistake. While the code is working, the used string for base64 encoding (line 187) does not match the excel lookup file. that means the codes sent to analytics contain some other codes. thanks to this update the codes for the props/evars should now match the excel file for classification import.